create messages container
